% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_swissPalm_table_listcols}
\alias{convert_swissPalm_table_listcols}
\title{Generate list-columns where necessary in SwissPalm tables}
\usage{
convert_swissPalm_table_listcols(sp_tbl)
}
\description{
Generate list-columns where necessary in SwissPalm tables
}
\keyword{internal}
