% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_values}
\alias{dataset_values}
\title{Mappings for dataset.value in getSWISSpalmData()}
\format{
\subsection{\code{dataset_values}}{

A vector of numbers with corresponding names which correspond key-value pairs of dataset.value
}
}
\usage{
dataset_values
}
\description{
Mappings for dataset.value in getSWISSpalmData()
}
\keyword{datasets}
