% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{Mappings for dataset parameter in \code{swissPalm()}}
\format{
An object of class \code{character} of length 7.
}
\usage{
datasets
}
\description{
A character vector with corresponding names for each available value of
\code{dataset} in \code{swissPalm()}
}
\examples{
swisspalmR::datasets[1:5]
}
\keyword{datasets}
