% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Perform extra actions on loading swisspalmR}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Ensures function runs. Not used.}

\item{pkgname}{Ensures function runs. Not used.}
}
\description{
Perform extra actions on loading swisspalmR
}
