% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWISSpalmData.R
\name{getSWISSpalmData}
\alias{getSWISSpalmData}
\title{Get palmitoylation data from SWISSpalm}
\usage{
getSWISSpalmData(
  protein.identifiers,
  download_dir = tempdir(),
  dataset.value = 1,
  species.value = 0
)
}
\arguments{
\item{protein.identifiers}{Character vector with protein identifiers. Should be an ID type supported by SWISSpalm (https://swisspalm.org/file_formats).}

\item{download_dir}{Directory in which temporary files will be downloaded. Defaults to R temporary directory.}

\item{dataset.value}{Which dataset should SWISSpalm use? Key-value pairs are in [swisspalm::dataset_values]. Default = 1 (all datasets)}

\item{species.value}{Which species should SWISSpalm search for? Key-value pairs for species are in [swisspalm::species_values]. Default = 0 (all species)}
}
\value{
Data frame containing palmitoylation data for proteins in protein.identifiers.
}
\description{
Get palmitoylation data from SWISSpalm
}
\references{
SwissPalm: Protein Palmitoylation database. Mathieu Blanc*, Fabrice P.A. David*, Laurence Abrami, Daniel Migliozzi, Florence Armand, Jérôme Burgi and F. Gisou van der Goot. F1000Research. (https://doi.org/10.12688/f1000research.6464.1)
}
