% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWISSpalmData.R
\name{getSWISSpalmData_httr}
\alias{getSWISSpalmData_httr}
\title{Get palmitoylation data from SWISSpalm using httr}
\usage{
getSWISSpalmData_httr(
  protein.identifiers,
  download_dir = tempdir(),
  dataset = "all",
  species = NA,
  verbose = F
)
}
\arguments{
\item{protein.identifiers}{Character vector with protein identifiers. Should be an ID type supported by SWISSpalm (https://swisspalm.org/file_formats).}

\item{download_dir}{Directory in which temporary files will be downloaded. Defaults to R temporary directory.}

\item{dataset}{Which dataset should SWISSpalm use? See valid values in [swisspalm::dataset_values]. Default = \code{"all"} (all datasets).}

\item{species}{Which species should SWISSpalm search for? See [swisspalm::species_values] for valid values. Default = 0 (all species)}

\item{verbose}{If TRUE, send status messages to the console.}
}
\value{
Data frame containing palmitoylation data for proteins in protein.identifiers.
}
\description{
Get palmitoylation data from SWISSpalm using httr
}
