% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseResponse}
\alias{parseResponse}
\alias{parseSWISSpalmResponse}
\title{Parse SWISSpalm response}
\usage{
parseSWISSpalmResponse(swiss.response)
}
\arguments{
\item{swiss.response}{Response from POST request to SWISSpalm}
}
\value{
Returns adequately parsed SWISSpalm response.
}
\description{
Parse/extract information from SWISSpalm
}
\author{
Simon Parker  \email{tadeo5@hotmail.co.uk}
}
