% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sendToSWISSpalm}
\alias{sendToSWISSpalm}
\title{Send to SWISSpalm}
\usage{
sendToSWISSpalm(
  identifiers,
  dataset.int = 1,
  species.int = 1,
  output.type = "text",
  output.dir
)
}
\arguments{
\item{identifiers}{foo1}

\item{output.type}{foo2}

\item{output.dir}{foo3}
}
\value{
Saves SWISSpalm output in specified user-specified output directory.
}
\description{
Send protein identifiers to SWISSpalm
}
\author{
Simon Parker  \email{tadeo5@hotmail.co.uk}
}
