% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Mappings for \code{species} parameter in \code{swissPalm()}}
\description{
A 92-length character vector with corresponding names for each available value of
\code{species} in \code{swissPalm()}.
}
\examples{
swisspalmR::species[1:5]
}
