% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_values}
\alias{species_values}
\title{Mappings for species.value parameter in getSWISSpalmData()}
\format{
\subsection{\code{dataset_values}}{

A vector of numbers with corresponding names which correspond key-value pairs of species.value
}
}
\usage{
species_values
}
\description{
Mappings for species.value parameter in getSWISSpalmData()
}
\keyword{datasets}
