% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissPalm.R
\name{swissPalm}
\alias{swissPalm}
\title{Get palmitoylation data from the swissPalm database}
\usage{
swissPalm(
  protein_id,
  download_dir = tempdir(),
  dataset = "all",
  species = NULL
)
}
\arguments{
\item{protein_id}{Character vector with protein identifiers. Should be an ID type supported by SWISSpalm (https://swisspalm.org/file_formats).}

\item{download_dir}{Directory in which temporary files will be downloaded. Default = R \code{tempdir()} command.}

\item{dataset}{Which dataset should SWISSpalm use? See valid values in [swisspalm::dataset_values]. Default = \code{"all"} (all datasets).}

\item{species}{Which species should SWISSpalm search for? See [swisspalm::species_values] for valid values. Default = \code{0} (all species).}

\item{verbose}{If TRUE, send status messages to the console.}
}
\value{
Data frame containing palmitoylation data for proteins in protein.identifiers.
}
\description{
Get palmitoylation data from the swissPalm database
}
