% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swisspalm_GET_req}
\alias{swisspalm_GET_req}
\title{Construct an HTTP GET request for SwissPalm}
\usage{
swisspalm_GET_req(user)
}
\arguments{
\item{user}{User agent for the HTTP request.}
}
\description{
Construct an HTTP GET request for SwissPalm
}
\keyword{internal}
