% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swisspalm_POST_req}
\alias{swisspalm_POST_req}
\title{Construct an HTTP POST request for SwissPalm}
\usage{
swisspalm_POST_req(headers, xhr, user)
}
\arguments{
\item{headers}{A named list of headers and their values for the POST request.}

\item{xhr}{A list of XHR components and their values for the POST request.}

\item{user}{User agent for the HTTP request.}
}
\description{
Construct an HTTP POST request for SwissPalm
}
\keyword{internal}
