% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swisspalmR-package.R
\docType{package}
\name{swisspalmR-package}
\alias{swisspalmR}
\alias{swisspalmR-package}
\title{swisspalmR: Access SwissPalm data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access 'SwissPalm' data through HTTP
}
\author{
\strong{Maintainer}: Simon Parker \email{simon.parker@lshtm.ac.uk} (\href{https://orcid.org/0009-0003-8214-4496}{ORCID}) [copyright holder]

}
\keyword{internal}
