% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_values}
\alias{dataset_values}
\title{Mappings for dataset in getSWISSpalmData()}
\format{
\subsection{\code{dataset_values}}{

A character vector with corresponding names for each available value of
\code{dataset} in \code{getSWISSpalmData()}
}
}
\usage{
dataset_values
}
\description{
Mappings for dataset in getSWISSpalmData()
}
\keyword{datasets}
