% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{Mappings for \code{dataset} parameter in \code{swissPalm()}}
\description{
A 7-length character vector with corresponding names for each available value of
\code{dataset} in \code{swissPalm()}
}
\examples{
swisspalmR::datasets[1:5]
}
