% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sendToSWISSpalm}
\alias{sendToSWISSpalm}
\title{Send to SWISSpalm}
\usage{
sendToSWISSpalm(identifiers, dataset.value = "all", species.value = 1)
}
\arguments{
\item{identifiers}{Vector of protein identifiers.}

\item{dataset.value}{Value indicating the desired dataset.}

\item{species.value}{Value indicating the species preference.}

\item{output.dir}{If specified, indicates the directory in which the results will be saved.}
}
\value{
Returns parsed SWISSpalm output.
}
\description{
Send protein identifiers to SWISSpalm
}
\author{
Simon Parker  \email{tadeo5@hotmail.co.uk}
}
