% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Mappings for species parameter in \code{swissPalm()}}
\format{
An object of class \code{character} of length 92.
}
\usage{
species
}
\description{
A character vector with corresponding names for each available value of
\code{species} in \code{swissPalm()}.
}
\examples{
swisspalmR::species[1:5]
}
\keyword{datasets}
