% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_values}
\alias{species_values}
\title{Mappings for species parameter in getSWISSpalmData()}
\format{
\subsection{\code{species_values}}{

A character vector with corresponding names for each available value of
\code{species} in \code{getSWISSpalmData()}
}
}
\usage{
species_values
}
\description{
Mappings for species parameter in getSWISSpalmData()
}
\keyword{datasets}
